pyproject.toml
[tool.poetry]
name = "michaelscottbot"
version = "0.2.0"
description = "A Discord bot that shares Michael Scott quotes"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^2.3.2"
flask = "^3.0.0"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
